[Fix] Move createBucketlist api call to helpers file
