+++ src/components/BucketlistCard.js
@@ -18,7 +18,7 @@ class BucketlistCard extends Component {
         this.setState({ selected_bucketlist: this.props.bucketlist.id });
     }
 
-    
+
 
     render() {
         return (
@@ -32,16 +32,13 @@ class BucketlistCard extends Component {
                             id: this.props.bucketlist.id,
                         }}
                         className="bucketlist-name"
-                     name={this.props.bucketlist.name}>
+                        name={this.props.bucketlist.name}>
                         <CardTitle title={this.props.bucketlist.name} />
                     </Link>
-
-
 
                     <CardText style={{ "fontSize": 1.2 + "em" }}>
                         {this.props.bucketlist.description}
                     </CardText>
-
 
                     <CardActions>
                         <FlatButton label="Update" style={{ width: 40 + "%" }}>
@@ -57,7 +54,6 @@ class BucketlistCard extends Component {
             </div>
         )
     }
-
 }
 
 export default BucketlistCard;
