+++ src/components/BucketlistsView.js
@@ -4,6 +4,7 @@ import ROOT_URL from '../constants';
 import axios from 'axios';
 import BucketlistCard from './BucketlistCard';
 import InNavAppBar from '../navigation/InNavAppBar';
+import helpers from '../helpers';
 
 
 class BucketlistsView extends Component {
@@ -23,28 +24,6 @@ class BucketlistsView extends Component {
 
   componentWillReceiveProps(nextProps) {
     this.setState({ bucketlists: nextProps.bucketlists });
-  }
-
-  createBucketlist = (e, createForm) => {
-    e.preventDefault();
-    let name = createForm.state.name
-    let description = createForm.state.description;
-    let payload = { "name": name, "description": description };
-
-    axios({
-      method: 'post',
-      url: ROOT_URL + '/api/v1/bucketlists',
-      data: payload,
-      headers: {
-        "Content-Type": "application/json",
-        "Authorization": `JWT ${window.sessionStorage.accessToken}`
-      }
-    }).then(function (response) {
-      this.setState({ name: payload.name, description: payload.description });
-      this.props.addBucketlist(payload);
-      this.props.getBucketlists();
-
-    }.bind(this));
   }
 
   handleFocus(e) {
@@ -77,7 +56,7 @@ class BucketlistsView extends Component {
     } else {
       bucketlists_list = bucketlists.map((bucketlist, index) => (
         <div className="box" key={bucketlist.id}>
-          <BucketlistCard bucketlist={bucketlist} getBucketlists={this.props.getBucketlists}/>
+          <BucketlistCard bucketlist={bucketlist} getBucketlists={this.props.getBucketlists} />
         </div>)
 
       );
@@ -85,22 +64,22 @@ class BucketlistsView extends Component {
 
     return (
       <div className="BucketlistsView">
-      <InNavAppBar />
+        <InNavAppBar />
         <h3 >My Bucketlists</h3>
         <div>
-        <input id="search" type="search"
-          className="search-field"
-          ref="search"
-          defaultValue="Search" required
-          onFocus={this.handleFocus.bind(this)}
-          onChange={this.handleChange.bind(this)} />
-          </div>
+          <input id="search" type="search"
+            className="search-field"
+            ref="search"
+            defaultValue="Search" required
+            onFocus={this.handleFocus.bind(this)}
+            onChange={this.handleChange.bind(this)} />
+        </div>
         <div className="wrapper">
           {bucketlists_list}
         </div>
         <CreateBucketlistForm bucketlists={this.props.bucketlists}
           addBucketlist={this.props.addBucketlist.bind(this)}
-          createBucketlist={this.createBucketlist.bind(this)} />
+          createBucketlist={helpers.createBucketlist.bind(this)} />
       </div>
     );
   }
